face_api/requirements.txt
sklearn==0.0
subprocess32==3.2.7
tensorboard==1.6.0
tensorflow==1.6.0
djangorestframework==3.6.4
django-rest-swagger==2.1.2
dlib==19.9.0
Django==1.11.2
django-cors-headers==2.2.0
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.4
numpy==1.14.2
scipy
python-dateutil
imutils
